Convert Class Movie in abstract class
