+++ task1/src/main/java/com/fundacionjala/movies/Rental.java
@@ -1,22 +1,22 @@
 package com.fundacionjala.movies;
 
-public abstract class Rental {
+public  class Rental {
 
     private final Movie movie;
 
     private final int daysRented;
-    public static final int DEFAULT_FREQUENT_POINT = 1;
-
 
     public Rental(Movie movie, int daysRented) {
         this.movie = movie;
         this.daysRented = daysRented;
     }
 
-    public abstract double calculateCharge();
+    public double calculateCharge() {
+        return movie.calculateChargeMovie(daysRented);
+    }
 
     public int calculateFrequentRenterPoints() {
-        return DEFAULT_FREQUENT_POINT;
+        return movie.calculateFrequentPointsMovie(daysRented);
     }
 
     public int getDaysRented() {
